reduce indent merge small files
